api/api/requirements.txt
click==8.1.7
colorama==0.4.6
Flask==2.0.1
Flask-Login==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.3
greenlet==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
psycopg2-binary==2.9.9
SQLAlchemy==1.4.32
Werkzeug==2.0.3
WTForms==3.1.2
